Move fformat package inside sfa folder
